@@ -7,17 +7,19 @@ import io.joshworks.fstore.log.segment.block.FixedSizeEntryBlock;
 
 import java.nio.ByteBuffer;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class FixedSizeBlockSerializer<T> implements Serializer<FixedSizeEntryBlock<T>> {
 
     private final Codec codec;
-    private Serializer<T> serializer;
+    private final Serializer<T> serializer;
     private final int entrySize;
 
-    FixedSizeBlockSerializer(Serializer<T> serializer, int entrySize) {
-        this.codec = new SnappyCodec();
+    FixedSizeBlockSerializer(Serializer<T> serializer, int entrySize, boolean compress) {
+        this.codec = compress ? new SnappyCodec() : Codec.noCompression();
         this.serializer = serializer;
         this.entrySize = entrySize;
     }
@@ -32,10 +34,17 @@ public class FixedSizeBlockSerializer<T> implements Serializer<FixedSizeEntryBlo
         //do nothing
     }
 
+    private static final Map<Integer, List<Integer>> cachedLengths = new ConcurrentHashMap<>();
+
     @Override
     public FixedSizeEntryBlock<T> fromBytes(ByteBuffer buffer) {
         ByteBuffer data = codec.decompress(buffer);
-        List<Integer> lengths = IntStream.range(0, data.limit() / entrySize).boxed().map(i -> entrySize).collect(Collectors.toList());
+
+        int entries = data.limit() / entrySize;
+        //FIXME concurrent modification exception
+        List<Integer> lengths = cachedLengths.computeIfAbsent(entries, k -> IntStream.range(0, data.limit() / entrySize).boxed().map(i -> entrySize).collect(Collectors.toList()));
+
         return new FixedSizeEntryBlock<>(serializer, lengths, data, entrySize);
     }
+
 }
